Type casts to shut up compiler warnings

+++ trunk/src/gmt_cdf.c
@@ -1,5 +1,5 @@
 /*--------------------------------------------------------------------
- *	$Id: gmt_cdf.c,v 1.57 2010-03-22 18:55:44 guru Exp $
+ *	$Id: gmt_cdf.c,v 1.58 2010-03-23 00:05:42 jluis Exp $
  *
  *	Copyright (c) 1991-2010 by P. Wessel and W. H. F. Smith
  *	See LICENSE.TXT file for copying and redistribution conditions.
@@ -277,8 +277,8 @@ GMT_LONG GMT_cdf_read_grd (struct GRD_HEADER *header, float *grid, double w, dou
 		}
 	}
 
-	header->nx = width_in;
-	header->ny = height_in;
+	header->nx = (int)width_in;
+	header->ny = (int)height_in;
 	header->x_min = w;
 	header->x_max = e;
 	header->y_min = s;
@@ -350,8 +350,8 @@ GMT_LONG GMT_cdf_write_grd (struct GRD_HEADER *header, float *grid, double w, do
 	header->x_max = e;
 	header->y_min = s;
 	header->y_max = n;
-	header->nx = width_out;
-	header->ny = height_out;
+	header->nx = (int)width_out;
+	header->ny = (int)height_out;
 
 	/* Write grid header */
